Change optimiser start point
Results of tests in issue #5 showed that the best strategy is to use bfgs for a few iterations, in cases where it converges this gives a quicker fit. It will converge after around 150 function evals but is slower than Newton-Raphson in that case. Adding a better start point helps a bit