Start to add fastlmm option
Haven't tested, but most of code is there. Printing is different by model, and hasn't yet been implemented.

Streams one SNP at a time rather than in large blocks, which should be tested before proceeding as may be very inefficient.

If this way works, there is a lot of redundancy in model.py that should be removed before finalising. i.e. combine lmm/binary/continuous and use a different kiter to pick fit