Check covariates samples matching with LMM
If the covariate file contains a higher number of samples than the
phenotype file, the covariates are ignored with no warning given to the
user. This change actually check that the sample names are all matching
and stops the execution if it's not so. Because of the LMM cache
saving/loading it is not desirable to reduce the covariates matrix at
runtime. Better to fail than run with unexepcted behavior.

tentatively closes #110